backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "companionhk-backend"
version = "0.1.0"
description = "CompanionHK FastAPI orchestration service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic",
  "fastapi",
  "langchain-core",
  "langchain-openai",
  "langgraph",
  "pgvector",
  "python-multipart",
  "pydantic-settings",
  "psycopg[binary]",
  "redis",
  "requests",
  "sqlalchemy",
  "uvicorn[standard]",
]

[project.optional-dependencies]
dev = ["httpx", "pytest", "pytest-asyncio", "ruff"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["app"]
